requirements.txt
blinker==1.4
flask-cors==3.0.2
flask-login==0.4.0
flask-oauthlib==0.9.3
flask-restful==0.3.5
flask-wtf==0.14.2
flask==0.12
pyjwt==1.4.2
unidecode==0.4.20
webargs==1.5.3
google-api-python-client==1.6.2
